+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next version
 
 - Add `Agilibox::MiniFormObject`
+- Replace time period filter by date+datetime period filters
 
 ## 1.1.0